Move around more, rename to readonly
